refactor(engine): remove reth-engine-service crate
EngineService was a thin wrapper around ChainOrchestrator that added no
behavior — just wiring + a Stream delegation. Move the construction
logic into reth-engine-tree::launch::build_engine_orchestrator, which
returns ChainOrchestrator directly.

Amp-Thread-ID: https://ampcode.com/threads/T-019c5c71-6817-72c4-80c9-3e7dc480088a
