--- crates/engine/service/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "reth-engine-service"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-# reth
-reth-consensus.workspace = true
-reth-engine-tree.workspace = true
-reth-evm.workspace = true
-reth-network-p2p.workspace = true
-reth-payload-builder.workspace = true
-reth-provider.workspace = true
-reth-prune.workspace = true
-reth-stages-api.workspace = true
-reth-tasks.workspace = true
-reth-node-types.workspace = true
-reth-chainspec.workspace = true
-reth-engine-primitives.workspace = true
-reth-trie-db.workspace = true
-
-# async
-futures.workspace = true
-pin-project.workspace = true
-
-# misc
-
-[dev-dependencies]
-reth-engine-tree = { workspace = true, features = ["test-utils"] }
-reth-ethereum-consensus.workspace = true
-reth-ethereum-engine-primitives.workspace = true
-reth-evm-ethereum.workspace = true
-reth-exex-types.workspace = true
-reth-primitives-traits.workspace = true
-reth-node-ethereum.workspace = true
-reth-trie-db.workspace = true
-
-alloy-eips.workspace = true
-tokio = { workspace = true, features = ["sync"] }
-tokio-stream.workspace = true
